WIP: Refactoring packages to prevent import loop

+++ matches/matches.go
@@ -12,13 +12,16 @@ package matches
 import (
 	"encoding/csv"
 	"fmt"
+	"io/ioutil"
 	"log"
 	"os"
+	"path/filepath"
 	"sort"
 	"strconv"
 	"text/tabwriter"
 
 	"github.com/atc0005/bridge/checksums"
+	"github.com/atc0005/bridge/paths"
 	"github.com/atc0005/bridge/units"
 
 	"github.com/360EntSecGroup-Skylar/excelize"
@@ -266,6 +269,120 @@ func (fm FileMatch) GenerateCSVDataRow() []string {
 	}
 }
 
+// NewFileSizeIndex optionally recursively processes a provided path and returns a
+// slice of FileMatch objects
+func NewFileSizeIndex(recursiveSearch bool, ignoreErrors bool, fileSizeThreshold int64, path string) (FileSizeIndex, error) {
+
+	fileSizeIndex := make(FileSizeIndex)
+	var err error
+
+	//log.Println("RecursiveSearch:", recursiveSearch)
+
+	if recursiveSearch {
+
+		// Walk walks the file tree rooted at path, calling the anonymous function
+		// for each file or directory in the tree, including path. All errors that
+		// arise visiting files and directories are filtered by the anonymous
+		// function. The files are walked in lexical order, which makes the output
+		// deterministic but means that for very large directories Walk can be
+		// inefficient. Walk does not follow symbolic links.
+		err = filepath.Walk(path, func(path string, info os.FileInfo, err error) error {
+
+			// If an error is received, check to see whether we should ignore
+			// it or return it. If we return a non-nil error, this will stop
+			// the filepath.Walk() function from continuing to walk the path,
+			// and your main function will immediately move to the next line.
+			if err != nil {
+				if !ignoreErrors {
+					return err
+				}
+
+				// WARN
+				log.Println("Error encountered:", err)
+				log.Println("Ignoring error as requested")
+
+			}
+
+			// make sure we're not working with the root directory itself
+			if path != "." {
+
+				// ignore directories
+				if info.IsDir() {
+					return nil
+				}
+
+				// ignore files below the size threshold
+				if info.Size() < fileSizeThreshold {
+					return nil
+				}
+
+				// If we made it to this point, then we must assume that the file
+				// has met all criteria to be evaluated by this application.
+				// Let's add the file to our slice of files of the same size
+				// using our index based on file size.
+				fileSizeIndex[info.Size()] = append(
+					fileSizeIndex[info.Size()],
+					FileMatch{
+						FileInfo:        info,
+						FullPath:        path,
+						ParentDirectory: filepath.Dir(path),
+					})
+			}
+
+			return err
+		})
+
+	} else {
+
+		// If recursiveSearch is not enabled, process just the provided path
+
+		// err is already declared earlier at a higher scope, so do not
+		// redeclare here
+		var files []os.FileInfo
+		files, err = ioutil.ReadDir(path)
+
+		if err != nil {
+			// TODO: Wrap error?
+			log.Printf("Error from ioutil.ReadDir(): %s", err)
+
+			return fileSizeIndex, err
+		}
+
+		// Use []os.FileInfo returned from ioutil.ReadDir() to build slice of
+		// FileMatch objects
+		for _, file := range files {
+
+			// ignore directories
+			if file.IsDir() {
+				continue
+			}
+
+			// ignore files below the size threshold
+			if file.Size() < fileSizeThreshold {
+				continue
+			}
+
+			// If we made it to this point, then we must assume that the file
+			// has met all criteria to be evaluated by this application.
+			// Let's add the file to our slice of files of the same size
+			// using our index based on file size.
+			fileSizeIndex[file.Size()] = append(
+				fileSizeIndex[file.Size()],
+				FileMatch{
+					FileInfo: file,
+					FullPath: filepath.Join(path, file.Name()),
+					// ParentDirectory: filepath.Dir(path),
+					// `path` is a flat directory structure (we are not using
+					// recursion), so record it directly as the parent
+					// directory for files within
+					ParentDirectory: path,
+				})
+		}
+	}
+
+	return fileSizeIndex, err
+}
+
 // PruneFileSizeIndex removes map entries with single-entry slices which do
 // not reflect potential duplicate files (i.e., duplicate file size !=
 // duplicate files)
@@ -453,6 +570,10 @@ func (fi FileChecksumIndex) WriteFileMatchesWorkbook(filename string, summary Du
 // WriteFileMatchesCSV writes duplicate files recorded in a FileChecksumIndex
 // to the specified CSV file.
 func (fi FileChecksumIndex) WriteFileMatchesCSV(filename string) error {
+
+	if !paths.PathExists(filepath.Dir(filename)) {
+		return fmt.Errorf("parent directory for specified CSV file to create does not exist")
+	}
 
 	file, err := os.Create(filename)
 	if err != nil {
